start refactoring references to module
